Main Ui
Main UI, settings, quit
