--- Assets/Scripts/UI/CombineButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 438c31e5ddd11e540bf8048a66b32cd1
