--- Assets/Scripts/UI/ElementSelector.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 74e7b34631d0c1c4dbd1b247662c4cf5
