+++ Assets/Scripts/UI/Main Menu UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 75c774644b837f849b2de7f22ce6c68b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
