+++ Assets/Scripts/UI/Main Menu UI/Main Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f3266be2807b1c4baf16bbb264f0a59
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
