+++ Assets/Scripts/UI/Main Menu UI/Main Menu/SettingsUI.cs
@@ -1,0 +1,26 @@
+using System.Collections;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class SettingsUI : MonoBehaviour
+{
+    private Animator animator;
+
+    private void Awake()
+    {
+        animator = GetComponent<Animator>();
+    }
+
+    public void Close()
+    {
+        StartCoroutine(CloseAfterDelay());
+    }
+
+    private IEnumerator CloseAfterDelay()
+    {
+        animator.SetTrigger("close");
+        yield return new WaitForSeconds(0.5f);
+        gameObject.SetActive(false);
+        animator.ResetTrigger("close");
+    }
+}
