+++ Assets/TextMesh Pro/Fonts/NotoSansKR-VariableFont_wght SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 21a9650fac53c5b48b5e886d53b61ab1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
