+++ Assets/UI/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d79ed2b75e904b42a72b76cf810340a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
