+++ Assets/UI/Animations/Animation SettingsUI.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68d0773fdf2e30248b46ff6d99b7fd7b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
