+++ Assets/UI/Animations/Seetings UI.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: feac90e7df508144fbc853e2ab6feec6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
